Using assembly image to generate jp2

+++ lib/was/thumbnail_service/capture/capture_thumbnail.rb
@@ -8,12 +8,12 @@ module Was
           @druid_id = druid_id
           @memento_uri = memento_uri
           @memento_datetime = memento_datetime
-          @temporary_file = "#{Rails.configuration.thumbnail_tmp_directory}/#{@memento_datetime}.jpg"
+          @temporary_file = "#{Rails.configuration.thumbnail_tmp_directory}/#{@memento_datetime}"
         end
         
         def process_thumbnail
           result = capture 
-          if result.length > 0 then
+          if result.length > 0 && result.starts_with('#FAIL#') then
             if File.exist?(@temporary_file) then File.delete(@temporary_file) end
             raise "Thumbnail for memento #{@memento_uri} can't be generated."
           end
@@ -24,7 +24,11 @@ module Was
         def capture
           result = ""
           begin
-            result = Phantomjs.run(Rails.configuration.phantom_js_script, @memento_uri, @temporary_file)
+            result = Phantomjs.run(Rails.configuration.phantom_js_script, @memento_uri, @temporary_file+".jpeg")
+            if Rails.configuration.jp2_required then
+              Assembly::Image.new(@temporary_file+'.jpeg').create_jp2(:output=>@temporary_file+".jp2")
+              FileUtils.rm @temporary_file+'.jpeg'
+            end
           rescue Exception => e 
             result = result +"\nException in generating thumbnail. #{e.message}\n#{e.backtrace.inspect}"
           end
@@ -37,10 +41,14 @@ module Was
         end
         
         def save_to_stack
-          puts @druid_id
           stacks_dir = DruidTools::StacksDruid.new(@druid_id, Rails.configuration.image_stacks).content_dir
-          thumbnail_file = "#{stacks_dir}/#{@memento_datetime}.jpg"
-          FileUtils.mv @temporary_file, thumbnail_file
+          if Rails.configuration.jp2_required then
+            thumbnail_file = "#{stacks_dir}/#{@memento_datetime}.jp2"
+            FileUtils.mv @temporary_file+'.jp2', thumbnail_file
+          else
+            thumbnail_file = "#{stacks_dir}/#{@memento_datetime}.jpeg"
+            FileUtils.mv @temporary_file+".jpeg", thumbnail_file
+          end
         end
       end
     end
